refactor: Extract PlayerBar into composable hooks and components
Break the monolithic PlayerBar (~1700 lines) into focused, testable modules:
- useDualDeckAudio: dual-deck audio element management
- useCrossfade: equal power crossfade between decks
- useStallRecovery: stall detection and automatic recovery
- useMediaSession: iOS/Android lock screen media controls
- usePlayerKeyboardShortcuts: keyboard shortcut handling
- usePlaybackStateSync: store/audio state synchronization with
  WiFi reconnect recovery (wasPlayingBeforeUnload persistence)
- AlbumArt: extracted album art component with Navidrome fallback
